requirements.txt
requests[socks]
lxml
beautifulsoup4
pymysql
soupsieve
discord-webhook
json
opencv-python
pillow
photohash
numpy
matplotlib